Chapter 8

+++ src/main/java/mypackage/entity/ProductSuggestion.java
@@ -20,5 +20,5 @@ public class ProductSuggestion {
 
     private Double recommendedPrice;
 
-    private Double estimatedAnnualSales;
+    private Long estimatedAnnualSales;
 }
